RS-14 refactor(games): fix according to comments
