+++ src/games/sprint/constants.ts
@@ -1,0 +1,2 @@
+export const TIMER_NINE = 9;
+export const TIMER_ZERO = 0;
