+++ src/games/sprint/startTimer.ts
@@ -1,26 +1,26 @@
 import sprintStore from '../../store/sprintStore';
 import { clearAndGetElement } from '../../utils';
-import renderSprintResult from './renderSprintResult';
+import { TIMER_NINE, TIMER_ZERO } from './constants';
+import renderSprintResult from './results/renderSprintResult';
 
 const startTimer = (time: number, selector: string) => {
   const timerEl = document.querySelector(selector) as HTMLDivElement;
-  let t = time;
+  let currentTime = time;
   function timer() {
-    timerEl.textContent = String(t);
-    t -= 1;
-    if (t < 9) {
+    timerEl.textContent = String(currentTime);
+    currentTime -= 1;
+    if (currentTime < TIMER_NINE) {
       const addZero = timerEl.textContent;
       timerEl.textContent = `0${addZero}`;
     }
-    if (t < 0) {
+    if (currentTime < TIMER_ZERO) {
       clearInterval(sprintStore.timerId);
-      timerEl.textContent = 'Time Off';
       timerEl.remove();
       const gamesContent = clearAndGetElement('.games__content') as HTMLDivElement;
       gamesContent.append(renderSprintResult(startTimer));
     }
   }
-  sprintStore.timerId = setInterval(timer, 1000);
+  sprintStore.timerId = window.setInterval(timer, 1000);
 };
 
 export default startTimer;
