+++ src/sass/components/_footer.scss
@@ -1,4 +1,5 @@
 .footer {
+  min-height: 90px;
   padding: 25px 0;
   margin-top: 20px;
   border-top: 1px solid rgba(0 0 0 / 12.5%);
