+++ src/sass/pages/games/_games.scss
@@ -1,5 +1,5 @@
 .games {
-  height: 100vh;
+  height: calc(100vh - 180px);
 
   &__wrapper {
     display: flex;
