+++ src/utils.ts
@@ -8,7 +8,10 @@ export const clearAndGetElement = (selector: string) => {
 
 export const shuffleArray = (array: []) => {
   const newArray = array.slice();
-  newArray.sort(() => Math.random() - 0.5);
+  for (let i = newArray.length - 1; i > 0; i -= 1) {
+    const j = Math.floor(Math.random() * (i + 1));
+    [newArray[i], newArray[j]] = [newArray[j], newArray[i]];
+  }
   return newArray;
 };
